Fixed possible index-out-of-bounds during shutdown and removed some unnecessary typecasting

+++ src/ui/artists_page/mod.rs
@@ -20,7 +20,7 @@ impl ArtistsPage {
     }
 
     #[inline]
-    pub fn assign_artwork(&self, index: u32, artwork: Option<gdk::Texture>) {
+    pub fn assign_artwork(&self, index: usize, artwork: Option<gdk::Texture>) {
         self.imp().assign_artwork(index, artwork);
     }
